// Loan application custom UI sample/spondulix-loan-app/src/app/providers/models/pfInstance.config.ts
/*
* Copyright © 2021. TIBCO Software Inc.
* This file is subject to the license terms contained
* in the license file that is distributed with this file.
*/

import {PFEInstanceInfo} from '../models/instanceInfo';

export class PfInstanceConfig {
  processName: string;
  moduleName: string;
  moduleVersion: string;
  caseData: any;
  formalParameterName: string;
  caseRef: string;
}